requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-embed-video==1.4.0
django-js-asset==1.2.2
django-parler==2.2
django-parler-rest==2.1
django-rosetta==0.9.7
djangorestframework==3.12.4
fontawesome-free==5.15.4
future==0.18.2
html5lib==1.1
idna==3.2
importlib-metadata==3.10.1
Pillow==8.3.2
polib==1.1.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.6.0
